requirements-dev.txt
pytest
pyyaml
rhosocial-activerecord-testsuite
pytest-asyncio
pydantic[email]
pytest-timeout
